[restructuring] generalservice 위치 변경
